Move primary code to __init__.py
